requirements.txt
pysocks
beautifulsoup4>=4.7.1
requests>=2.21.0
yara